addressing nsthorat comments
