+++ src/math/webgl/sigmoid_gpu_test.ts
@@ -16,7 +16,7 @@ limitations under the License.
 import * as test_util from '../../test_util';
 import * as util from '../../util';
 import {UnaryOpProgram, UnaryOp} from './unaryop_gpu';
-import * as unaryop_gpu from './unaryop_gpu';
+import * as unaryop_gpu_test from './unaryop_gpu_test';
 import {Array2D} from '../ndarray';
 
 describe('sigmoid_gpu', () => {
@@ -42,5 +42,5 @@ describe('sigmoid_gpu', () => {
 function uploadSigmoidDownload(
     a: Float32Array, rows: number, cols: number): Float32Array {
   const arr = Array2D.new([rows, cols], a);
-  return unaryop_gpu.uploadUnaryDownload(arr, UnaryOp.SIGMOID);
+  return unaryop_gpu_test.uploadUnaryDownload(arr, UnaryOp.SIGMOID);
 }
